+++ mimesis/providers/development.py
@@ -10,9 +10,7 @@ from mimesis.data import (
     STAGES,
     SYSTEM_QUALITY_ATTRIBUTES,
 )
-from mimesis.enums import DSNType
 from mimesis.providers.base import BaseProvider
-from mimesis.providers.internet import Internet
 
 __all__ = ["Development"]
 
@@ -22,25 +20,9 @@ class Development(BaseProvider):
 
     def __init__(self, *args: t.Any, **kwargs: t.Any) -> None:
         super().__init__(*args, **kwargs)
-        self._internet = Internet(
-            random=self.random,
-            *args,
-            **kwargs,
-        )
-        self.__now = datetime.now()
 
     class Meta:
         name = "development"
-
-    def dsn(self, dsn_type: DSNType | None = None, **kwargs: t.Any) -> str:
-        """Generates a random DSN (Data Source Name).
-
-        :param dsn_type: DSN type.
-        :param kwargs: Additional arguments for Internet.hostname().
-        """
-        hostname = self._internet.hostname(**kwargs)
-        scheme, port = self.validate_enum(dsn_type, DSNType)
-        return f"{scheme}://{hostname}:{port}"
 
     def software_license(self) -> str:
         """Generates a random software license.
@@ -60,7 +42,7 @@ class Development(BaseProvider):
         :Example:
             2016.11.08
         """
-        year = self.random.randint(2016, self.__now.year)
+        year = self.random.randint(2016, datetime.now().year)
         month = self.random.randint(1, 12)
         day = self.random.randint(1, 29)
         return f"{year}.{month}.{day}"