+++ tests/test_providers/test_internet.py
@@ -5,7 +5,7 @@ import pytest
 import validators
 
 from mimesis import Internet, data
-from mimesis.enums import MimeType, PortRange, TLDType, URLScheme
+from mimesis.enums import DSNType, MimeType, PortRange, TLDType, URLScheme
 from mimesis.exceptions import NonEnumerableError
 
 from . import patterns
@@ -18,6 +18,23 @@ class TestInternet:
 
     def test_str(self, net):
         assert re.match(patterns.PROVIDER_STR_REGEX, str(net))
+
+    @pytest.mark.parametrize(
+        "dsn_type",
+        [
+            DSNType.POSTGRES,
+            DSNType.MYSQL,
+            DSNType.MONGODB,
+            DSNType.REDIS,
+            DSNType.COUCHBASE,
+            DSNType.MEMCACHED,
+            DSNType.RABBITMQ,
+        ],
+    )
+    def test_dsn(self, net, dsn_type):
+        scheme, port = dsn_type.value
+        assert net.dsn(dsn_type=dsn_type).endswith(f":{port}")
+        assert net.dsn(dsn_type=dsn_type).startswith(f"{scheme}://")
 
     @pytest.mark.parametrize(
         "subdomains",
@@ -385,3 +402,6 @@ class TestSeededInternet:
 
     def test_public_dns(self, i1, i2):
         assert i1.public_dns() == i2.public_dns()
+
+    def test_dsn(self, i1, i2):
+        assert i1.dsn() == i2.dsn()